pyproject.toml
[project]
name = "ai-backend-api"
version = "0.1.0"
description = "AI应用后端接口，提供翻译和总结功能"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "httpx>=0.25.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]
